+++ bundlers/BundleBuilder.ts
@@ -2,9 +2,9 @@
 import exorcist = require("exorcist");
 import vinylfs = require("vinyl-fs");
 import vinylSourceStream = require("vinyl-source-stream");
-import Q = require("q");
 import BrowserifyHelper = require("./BrowserifyHelper");
 import TypeScriptHelper = require("./TypeScriptHelper");
+import TsBrowserify = require("./browser/TsBrowserify");
 
 type BrowserifyObject = browserify.BrowserifyObject;
 type BrowserifyTransform = BrowserifyHelper.BrowserifyTransform;
@@ -105,10 +105,10 @@ module BundleBuilder {
      * @param rebuilder the browserify plugin to use if 'bundleOpts.rebuild' is true (normally watchify)
      * @param bundleOpts options for how to compile the bundle, are used to build browserify and are also passed along to the compileBundle function
      * @param compileBundle a function which takes a bundler, options, paths, and a bundle stream creator and compiles the bundle
-     * @param [optsModifier] a optional function which can modify the Browserify and browserPack options before they are passed to the browserify constructor
+     * @param optsModifier optional, function which can modify the Browserify and browserPack options before they are passed to the browserify constructor
      */
     export function buildBundler(
-        browserify: browserify.BrowserifyConstructor,
+        browserBundler: { new(opts: browserify.Options): browserify.BrowserifyObject },
         rebuilder: (b: browserify.BrowserifyObject, opts?: any) => browserify.BrowserifyObject,
         bundleOpts: BundleOptions,
         compileBundle: BrowserifyCompileFunc,
@@ -166,7 +166,7 @@ module BundleBuilder {
                         };
                     };
                 }
-                var bundler = createBrowserify(browserify, rebuilder, optsRes, bundleOpts, paths);
+                var bundler = createBrowserify(browserBundler, rebuilder, optsRes, bundleOpts, paths);
                 var transforms = _createTransforms(bundler);
                 compileBundle(transforms, bundler, bundleOpts, paths.dstDir, _bundleSourceCreator, _listeners);
             },
@@ -176,14 +176,14 @@ module BundleBuilder {
 
 
     /** Sets up options and paths and creates a new Browserify instance
-     * @param browserify the browserify constructor to use
-     * @param rebuilder optional browserify plugin to use if 'bundleOpts.rebuild' is true
-     * @param customOpts custom browserify/browser-pack constructor options in addition to the 'bundleOpts' parameter already provided, can be null
-     * @param bundleOpts options used to help construct the browserify/browser-pack constructor options
+     * @param browserBundler the browser-bundler constructor to use
+     * @param rebuilder optional browser-bundler plugin to use if 'bundleOpts.rebuild' is true
+     * @param customOpts custom browser-bundler/browser-pack constructor options in addition to the 'bundleOpts' parameter already provided, can be null
+     * @param bundleOpts options used to help construct the browser-bundler/browser-pack constructor options
      * @param paths code input/output paths for the bundle compiler
      */
-    export function createBrowserify<R>(browserify: browserify.BrowserifyConstructor, rebuilder: (b: browserify.BrowserifyObject, opts?: any) => browserify.BrowserifyObject,
-            customOpts: browserify.Options & browserPack.Options, bundleOpts: BundleOptions, paths: CodePaths): BrowserifyObject {
+    export function createBrowserify<R>(browserBundler: { new(opts: browserify.Options): R }, rebuilder: (b: BrowserifyObject, opts?: any) => BrowserifyObject,
+            customOpts: browserify.Options & browserPack.Options, bundleOpts: BundleOptions, paths: CodePaths): R {
         // setup browserify/browser-pack options
         var defaultOpts: browserify.Options & browserPack.Options = {
             debug: bundleOpts.debug,
@@ -195,7 +195,7 @@ module BundleBuilder {
         // setup bundler options
         var plugins = bundleOpts.rebuild ? [rebuilder] : [];
         var bundlerOpts = BrowserifyHelper.createOptions(Object.assign(defaultOpts, paths), plugins);
-        return new browserify(bundlerOpts);
+        return new browserBundler(bundlerOpts);
     }
 
 }